Rename pypi package to lco_ingester, and rename top level package to match. Restructure packages a bit so that settings is inside lco_ingester so that import are more clear and a settings package isnt installed in a clients venv, refactor settings file a bit to make it cleaner
