+++ setup.py
@@ -7,13 +7,13 @@ with open(path.join(this_directory, 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
 
 setup(
-    name='ingester',
-    version='0.0.9',
+    name='lco-ingester',
+    version='0.0.10',
     description='Ingest frames into the LCO Archive',
     long_description=long_description,
     long_description_content_type='text/markdown',
     url='https://github.com/LCOGT/ingester',
-    packages=['ingester', 'ingester.utils', 'settings', 'scripts'],
+    packages=['lco_ingester', 'lco_ingester.utils', 'lco_ingester.settings', 'lco_ingester.scripts'],
     python_requires='>=3.5',
     classifiers=[
         'License :: OSI Approved :: GNU General Public License v3 (GPLv3)'
@@ -28,7 +28,7 @@ setup(
     ],
     entry_points={
         'console_scripts': [
-            'ingest_frame = scripts.ingest_frame:main',
+            'lco_ingest_frame = lco_ingester.scripts.ingest_frame:main',
         ]
     }
 )
